+++ app/src/main/AndroidManifest.xml
@@ -2,8 +2,8 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
           package="com.nicolasfanin.IUASampleApp">
 
-    <uses-permission android:name="android.permission.INTERNET" />
-    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
+    <uses-permission android:name="android.permission.INTERNET"/>
+    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE"/>
 
     <application
             android:allowBackup="true"
@@ -12,10 +12,7 @@
             android:roundIcon="@mipmap/ic_launcher_round"
             android:supportsRtl="true"
             android:theme="@style/AppTheme">
-        <activity
-                android:name=".activities.GuitHubRepoActivity"
-                android:label="@string/title_activity_guit_hub_repo">
-        </activity>
+
         <activity android:name=".activities.SplashActivity" android:screenOrientation="portrait">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
@@ -35,6 +32,9 @@
         </activity>
         <activity android:name=".activities.CreditCardActivity">
 
+        </activity>
+        <activity
+                android:name=".activities.GitHubRepoActivity">
         </activity>
     </application>
