+++ app/src/main/res/layout/activity_guit_hub_repo.xml
@@ -4,7 +4,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context="com.nicolasfanin.IUASampleApp.activities.GuitHubRepoActivity">
+        tools:context="com.nicolasfanin.IUASampleApp.activities.GitHubRepoActivity">
